Bugfix: Modified return type for ConvertBandwidth   1. Modified return type for ConvertBandwidth, it's shoudle be int64

+++ contivmodel/contivModel.go
@@ -3428,7 +3428,7 @@ func restoreNetprofile() error {
 	return nil
 }
 
-func ConvertBandwidth(bandwidth string) int {
+func ConvertBandwidth(bandwidth string) int64 {
 	var rate int
 	const (
 		kilobytes = 1024
